fix: refactor and naming
